Relocate the bodyClass util.
